% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{applyParameterAbbreviations}
\alias{applyParameterAbbreviations}
\title{Utility method to replace long parameter names with short abbreviations.
The change is made within the plate's plateData object, and the plate's paramAbbrs data field is set.}
\usage{
applyParameterAbbreviations(plate, abbrInfo)
}
\arguments{
\item{plate}{the plate on which to apply the abbreviations.}

\item{abbrInfo}{a list object with parameter name keys and corresponding abbreviations as values.}
}
\value{
returns a plate object with plateData that uses abbreviations as column names
}
\description{
Utility method to replace long parameter names with short abbreviations.
The change is made within the plate's plateData object, and the plate's paramAbbrs data field is set.
}
