% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{assessDataCompleteness}
\alias{assessDataCompleteness}
\title{Runs an analysis on the aso object to report on data coverage for each parameter.
The missing data report is entered into the aso object.}
\usage{
assessDataCompleteness(aso)
}
\value{
returns the aso object containing the missing data report in the dataCoverageTables slot
}
\description{
Runs an analysis on the aso object to report on data coverage for each parameter.
The missing data report is entered into the aso object.
}
