% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{clonePlate}
\alias{clonePlate}
\title{A utility method to make a copy of a plate object
May not be needed since assignments and parameter passing are copying values into a new copy.}
\usage{
clonePlate(plate)
}
\arguments{
\item{plate}{the plate to clone}
}
\value{
new plate copy
}
\description{
A utility method to make a copy of a plate object
May not be needed since assignments and parameter passing are copying values into a new copy.
}
