% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{encodeDiscreteParameters}
\alias{encodeDiscreteParameters}
\title{encodeDiscreteParameters}
\usage{
encodeDiscreteParameters(plate, parameter, textValues, numericValues)
}
\arguments{
\item{plate}{plate containing the data to encode}

\item{parameter}{parameter that requires encoding}

\item{textValues}{expected initial discreate values}

\item{numericValues}{replacement numeric values}
}
\value{
returns a copy of the plate having value substitutions for the discrete parameter
}
\description{
encodeDiscreteParameters
}
