% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{exportRandomForestResults}
\alias{exportRandomForestResults}
\title{Exports a colelction of random forest results into one file, along with Benchmark Dose.}
\usage{
exportRandomForestResults(aso)
}
\arguments{
\item{aso}{The aso object containing the RandomForest Results}
}
\description{
Exports a colelction of random forest results into one file, along with Benchmark Dose.
}
