% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{filterLowDataParameters}
\alias{filterLowDataParameters}
\title{filterLowDataParameters}
\usage{
filterLowDataParameters(
  plate,
  thresholdType = "well_count",
  missingDataThreshold = 10
)
}
\arguments{
\item{plate}{plate on which to filter parameters with low data representation}

\item{thresholdType}{one of 'well_count' or 'well_percentage'}

\item{missingDataThreshold}{}
}
\value{
a plate with updated plate data
}
\description{
filterLowDataParameters
}
