% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{filterParametersOnList}
\alias{filterParametersOnList}
\title{Reduces plate data to a set of parameters to keep.}
\usage{
filterParametersOnList(plate, paramsToKeep)
}
\arguments{
\item{plate}{a plate object containing plateData}

\item{paramsToKeep}{list of parameters to keep/maintain in the plateData, others will be filtered out.}
}
\value{
returns a new plate object with plateData only containing the specified parameteters.
}
\description{
Reduces plate data to a set of parameters to keep.
}
