% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{FLIPRDataMachineLearning}
\alias{FLIPRDataMachineLearning}
\title{generatees random forest predictions for FLIPRDatas}
\usage{
FLIPRDataMachineLearning(FLIPRData, masking = NULL)
}
\arguments{
\item{FLIPRData}{a FLIPRData object containing transformed data that's ready for analysis}

\item{masking}{list of wells to optionally mask}
}
\value{
creates a prediction heatmap and dot plot, returns a Heatmap object.
}
\description{
generatees random forest predictions for FLIPRDatas
}
