% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getColumnAbbrToImpute}
\alias{getColumnAbbrToImpute}
\title{helper function to determine which parameters should undergo imputation.}
\usage{
getColumnAbbrToImpute(aso)
}
\arguments{
\item{aso}{an aso object having plate data loaded ans parameters to control the imputation process}
}
\value{
a list object with two fields incicating which columns to impute for zeros or NAs,
}
\description{
helper function to determine which parameters should undergo imputation.
}
