% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{getEdgeWellTrimmedMatrixFromPlate}
\alias{getEdgeWellTrimmedMatrixFromPlate}
\title{getEdgeWellTrimmedMatrixFromPlate is a utility method to specifically exclude wells that are on the edge of a plate.
This is a very specific action for certain data sets where edge wells are not used due to conditions like evaporation.}
\usage{
getEdgeWellTrimmedMatrixFromPlate(plate)
}
\arguments{
\item{plate}{The plate containing the data to be trimmed}
}
\value{
A copy of the plate containing data for the remaining wells.
}
\description{
getEdgeWellTrimmedMatrixFromPlate is a utility method to specifically exclude wells that are on the edge of a plate.
This is a very specific action for certain data sets where edge wells are not used due to conditions like evaporation.
}
