% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{getPlateDataByWellSet}
\alias{getPlateDataByWellSet}
\title{retrieve well data based on a collection of specified wells.}
\usage{
getPlateDataByWellSet(plate, wellSet)
}
\arguments{
\item{plate}{a plate object from which to extract a subset of plateData}

\item{wellSet}{the collection of wells (well ids) to pull data for.}
}
\value{
returns a dataframe object with all parameter data for the specified wells.
}
\description{
retrieve well data based on a collection of specified wells.
}
