% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{getPlateDataForSample}
\alias{getPlateDataForSample}
\title{Utility method to extract the plate data from a plate, for a given sample name (set of data rows) and parameter name (data column).}
\usage{
getPlateDataForSample(plate, sampleName, parameter)
}
\arguments{
\item{plate}{the plate object from which to extract data.}

\item{sampleName}{the name of the sample for which to extract data. Names are specified in the input platemap annotation file.}

\item{parameter}{the parameter for which to pull data}
}
\value{
returns a data frame with well annotations and corresponding data
}
\description{
Utility method to extract the plate data from a plate, for a given sample name (set of data rows) and parameter name (data column).
}
