% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{getPlateDataForSampleAllParameters}
\alias{getPlateDataForSampleAllParameters}
\title{A utility method to report annotations and all data for a given sample}
\usage{
getPlateDataForSampleAllParameters(plate, sampleName)
}
\arguments{
\item{plate}{the plate object from which to pull data}

\item{sampleName}{the name of the sample (compound) for which to pull data.}
}
\value{
a data frame containing annotations and all parameter data for a given sample name
}
\description{
A utility method to report annotations and all data for a given sample
}
