% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{getPlateGgplotFromMatrix}
\alias{getPlateGgplotFromMatrix}
\title{Title}
\usage{
getPlateGgplotFromMatrix(
  plateFormatData,
  plotTitle = "",
  showRowColLabels = T,
  showRowTitle = F,
  showColTitle = F,
  showLegend = F,
  maxSatCount = 10,
  elementDividers = F,
  scaleColors = c("#e8eb59", "#0740fa"),
  showPlateName = F,
  xTitle = "col",
  yTitle = "row"
)
}
\arguments{
\item{plateFormatData}{a matrix object with dimensions and data that conforms to plate data.}

\item{plotTitle}{an optional plate heatmap title}

\item{showRowColLabels}{boolean to show or hid column labels}

\item{showRowTitle}{boolean to show or hide row title}

\item{showColTitle}{boolean to show or hide row title}

\item{showLegend}{boolean to show or hide colorbar legend}

\item{maxSatCount}{maximumn number of wells to allow color saturation in the plate heatmap. This sets color scale limits such that this number of wells saturate,
evenly at both ends of the plate's value range. This reduces the effect of outliers on the color scale limits.}

\item{elementDividers}{boolean to hide or show borders around heatmap elements}

\item{scaleColors}{a vector of colors that bound the color scale}

\item{showPlateName}{a boolean to show or hide the name of the plate over the heatmap}

\item{xTitle}{a title to show on the x-axis (if showRowColLabels is true)}

\item{yTitle}{a title to show on the y-axis (if showRowColLabels is true)}
}
\description{
Title
}
