% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{getPlateValueCutoffs}
\alias{getPlateValueCutoffs}
\title{getPlateValueCutoffs a utility method that returns an upper and lower bound of values in the input matrix
given the maxSaturationCount which defines how many data elements may exceed the bounds.
Note that the staturated (beyond bound values) are split at the low and high ends of the bounds.
That means that if 10 elements are allowed to 'saturate' (exceed bounds), 5 fall off the high end and 5 fall off the low end.}
\usage{
getPlateValueCutoffs(plateMatrix, maxSaturationCount = 10)
}
\arguments{
\item{plateMatrix}{input plate dimension matrix}

\item{maxSaturationCount}{the number of points (plate wells) to fall off the returned scale (half will be off high end, half of these off the low end)}
}
\value{
returns a vector with two numeric values, a low-end cutoff and a high-end cutoff, in that order
}
\description{
The utility of this method is to set reFLIPRDatanable color scale bounds for heatmaps.
}
