% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{getRowAndColumnNames}
\alias{getRowAndColumnNames}
\title{getRowAndColumnNames is a utility method that returns the plate row and column labels}
\usage{
getRowAndColumnNames(plateDim = 384)
}
\arguments{
\item{plateDim}{dimenstion of the plate}
}
\value{
returns a list with two entries with keys 'rows' and 'cols'
}
\description{
getRowAndColumnNames is a utility method that returns the plate row and column labels
}
