% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{harmonizeParametersAcrossPlates}
\alias{harmonizeParametersAcrossPlates}
\title{A utility method to harmonize parameters across plates. Simply makes sure all plate reads contain the same parameters.}
\usage{
harmonizeParametersAcrossPlates(plateSet)
}
\arguments{
\item{plateSet}{the plateset to harmonize}
}
\value{
a plateset object with various plate reads containing the same, consistent variables.
}
\description{
A utility method to harmonize parameters across plates. Simply makes sure all plate reads contain the same parameters.
}
