% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{imputeData}
\alias{imputeData}
\title{Performs missing data imputation, replacing missing data according to a supplied method name.
Note that this method is not currently used in typical pre-processing.}
\usage{
imputeData(FLIPRData)
}
\arguments{
\item{FLIPRData}{a FLIPRData object that has a platepair and processing parameters}
}
\value{
the FLIPRData object now having missing values imputed
}
\description{
Performs missing data imputation, replacing missing data according to a supplied method name.
Note that this method is not currently used in typical pre-processing.
}
