% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{initializeLog}
\alias{initializeLog}
\title{Starts a log file.}
\usage{
initializeLog(logLevel = "INFO")
}
\arguments{
\item{logLevel}{logger threshold to use while processing}
}
\description{
Starts a log file.
}
