% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{loadPlateMaps}
\alias{loadPlateMaps}
\title{This function loads plate annotations, plate maps as specified in the FLIPRData object after initializing with parameter file}
\usage{
loadPlateMaps(FLIPRData)
}
\arguments{
\item{FLIPRData}{the input FLIPRData file generated by readFLIPRDataParameterFile}
}
\value{
returns a FLIPRData object containing a plateset object with loaded plate data and plate map annotations.
}
\description{
This function loads plate annotations, plate maps as specified in the FLIPRData object after initializing with parameter file
}
