% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{loadPlates}
\alias{loadPlates}
\title{This method loads plate data according the the plate files specified within the CalFluxData parameter file.
One plate object is created for each input plate file and the plate pairs are stored within the returned CalFluxData object.}
\usage{
loadPlates(CalFluxData)
}
\arguments{
\item{CalFluxData}{the input CalFluxData file generated by readCalFluxDataParameterFile}
}
\value{
returns a CalFluxData object containing a plateset object with loaded plate data.
}
\description{
This method loads plate data according the the plate files specified within the CalFluxData parameter file.
One plate object is created for each input plate file and the plate pairs are stored within the returned CalFluxData object.
}
