% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{missingDataReport}
\alias{missingDataReport}
\title{missingDataReport reports on plate missing values}
\usage{
missingDataReport(plate, plateSize = 384)
}
\arguments{
\item{plate}{ASO plate}

\item{plateSize}{plate size, e.g. 384 (default)}
}
\value{
returns a dataframe with columns 'parameter', 'good_count', 'missing_count', 'na_count', 'zero_count'
}
\description{
missingDataReport reports on plate missing values
}
