% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{missingPlateDataChart}
\alias{missingPlateDataChart}
\title{missingPlateDataChart}
\usage{
missingPlateDataChart(plate, plateSize = 384, missingValLimit = 10)
}
\arguments{
\item{plate}{plate object containing data to evaluate}

\item{plateSize}{plate well count, e.g. 384 (default)}

\item{missingValLimit}{limits the plots to only return plots for parameters for many missing data points}
}
\value{
returns a ggplot facet_wrap plot, max 4 columns, only parameters with more missing data than missingValLimit
}
\description{
missingPlateDataChart
}
