% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_individual_plate}
\alias{parse_individual_plate}
\title{Internal function for parsing vector corresponding to a single plate}
\usage{
parse_individual_plate(plate)
}
\arguments{
\item{plate}{the plate statistic list}
}
\value{
parsed plate data
}
\description{
Internal function for parsing vector corresponding to a single plate
}
\author{
Andrew Patt
}
