% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{platePairMissingDataReport}
\alias{platePairMissingDataReport}
\title{platePairMissingDataReport}
\usage{
platePairMissingDataReport(plateset, platePair, plateSize = 384)
}
\arguments{
\item{plateset}{a plateset object containing at least two plates}

\item{platePair}{a vector of plate names to compare}

\item{plateSize}{plate well count, e.g. 384, 96..}
}
\value{
returns a dataframe that summarizes the union of missing data across two referenced plates
}
\description{
platePairMissingDataReport
}
