% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{processAsoData}
\alias{processAsoData}
\title{Wrapper method to process a data set possibly containing multiple file paires
processes data, writes file outputs, returns an ASO object with processed data.}
\usage{
processAsoData(asoParameterXlsxFilePath)
}
\arguments{
\item{asoParameterXlsxFilePath}{file path to param file}
}
\value{
processed ASO object
}
\description{
Wrapper method to process a data set possibly containing multiple file paires
processes data, writes file outputs, returns an ASO object with processed data.
}
