% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{processCalFluxData}
\alias{processCalFluxData}
\title{Wrapper method to process a data set possibly containing multiple file paires
processes data, writes file outputs, returns a CalFluxData object with processed data.}
\usage{
processCalFluxData(CalFluxDataParameterXlsxFilePath)
}
\arguments{
\item{CalFluxDataParameterXlsxFilePath}{file path to param file}
}
\value{
processed CalFluxData object
}
\description{
Wrapper method to process a data set possibly containing multiple file paires
processes data, writes file outputs, returns a CalFluxData object with processed data.
}
