% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{processData}
\alias{processData}
\title{This method is a convenience method that displays a file chooser for selecting a CalFluxData Paramter Excel file for processing.
The method will run data procesing and output diagnostic plots and tables.}
\usage{
processData()
}
\value{
Returns a CalFluxData object containing analysis parameters from the parameter file, input, refined and transformed plate data.
}
\description{
This method is a convenience method that displays a file chooser for selecting a CalFluxData Paramter Excel file for processing.
The method will run data procesing and output diagnostic plots and tables.
}
