% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{processFLIPRData}
\alias{processFLIPRData}
\title{Wrapper method to process a data set possibly containing multiple file paires
processes data, writes file outputs, returns a FLIPRData object with processed data.}
\usage{
processFLIPRData(FLIPRDataParameterXlsxFilePath, paired_analysis = TRUE)
}
\arguments{
\item{FLIPRDataParameterXlsxFilePath}{file path to param file}

\item{paired_analysis}{boolean specifying if plate pairs are used for baseline correction}
}
\value{
processed FLIPRData object
}
\description{
Wrapper method to process a data set possibly containing multiple file paires
processes data, writes file outputs, returns a FLIPRData object with processed data.
}
