% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_stemonix_data}
\alias{read_stemonix_data}
\title{Reads in data from stemonix platform and converts to plate class}
\usage{
read_stemonix_data(filename)
}
\arguments{
\item{filename}{(string) path to file to read in}
}
\value{
plate class object
}
\description{
Reads in data from stemonix platform and converts to plate class
}
\author{
Andrew Patt
}
