% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_stemonix_metadata}
\alias{read_stemonix_metadata}
\title{reads platemap information}
\usage{
read_stemonix_metadata(filename, plate)
}
\arguments{
\item{filename}{input plate map file name}

\item{plate}{an initialized plate object}
}
\value{
returns a plate with data loaded.
}
\description{
reads platemap information
}
\author{
Andrew Patt
}
