% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{readAsoParameterFile}
\alias{readAsoParameterFile}
\title{Reads an input ASO parameter Excel file and returns an aso object.}
\usage{
readAsoParameterFile(filename)
}
\arguments{
\item{filename}{excel aso parmeter file}
}
\value{
returns an aso object ready for analysis. The ASO file will not have plate data but rather just instructions on data
analysis and input data files. This is typically the first function to run to generate an initialized aso object.
}
\description{
Reads an input ASO parameter Excel file and returns an aso object.
}
