% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{readCalFluxDataParameterFile}
\alias{readCalFluxDataParameterFile}
\title{Reads an input CalFluxData parameter Excel file and returns a CalFluxData object.}
\usage{
readCalFluxDataParameterFile(filename)
}
\arguments{
\item{filename}{excel CalFluxData parmeter file}
}
\value{
returns a CalFluxData object ready for analysis. The CalFluxData file will not have plate data but rather just instructions on data
analysis and input data files. This is typically the first function to run to generate an initialized CalFluxData object.
}
\description{
Reads an input CalFluxData parameter Excel file and returns a CalFluxData object.
}
