% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{readFLIPRDataParameterFile}
\alias{readFLIPRDataParameterFile}
\title{Reads an input FLIPRData parameter Excel file and returns a FLIPRData object.}
\usage{
readFLIPRDataParameterFile(filename)
}
\arguments{
\item{filename}{excel FLIPRData parmeter file}
}
\value{
returns a FLIPRData object ready for analysis. The FLIPRData file will not have plate data but rather just instructions on data
analysis and input data files. This is typically the first function to run to generate an initialized FLIPRData object.
}
\description{
Reads an input FLIPRData parameter Excel file and returns a FLIPRData object.
}
