% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{replaceBlanksAndNAs}
\alias{replaceBlanksAndNAs}
\title{Utility method to replace blank and NA data with replacement values.}
\usage{
replaceBlanksAndNAs(plateSet, blankReplacements, naReplacements)
}
\arguments{
\item{plateSet}{the plateset object to work on}

\item{blankReplacements}{a list object with parameter abbreviation keys and replacement values}

\item{naRelacements}{a list object with parameter abbreviations keys and replacement values}
}
\value{
a plateset object with plate data replaced according to input values
}
\description{
Utility method to replace blank and NA data with replacement values.
}
