% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{runPairedTTest}
\alias{runPairedTTest}
\title{Performes a paired t-test to compare CalFluxData treated samples from the reference/control state.}
\usage{
runPairedTTest(CalFluxData)
}
\arguments{
\item{CalFluxData}{CalFluxData object having loaded plate data.}
}
\value{
a dataframe containing stat results for each parameter, for each well.
}
\description{
Performes a paired t-test to compare CalFluxData treated samples from the reference/control state.
}
