% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{runQcForReferencePlate}
\alias{runQcForReferencePlate}
\title{Runs quality control on the reference (control) plate read, prior to treatment.
This collects information on parameter variation (standard deviation, and coefficient of variation)
and z-scores for each parameter and each well help to identy outlier/bad wells that we might want to 'mask'}
\usage{
runQcForReferencePlate(CalFluxData)
}
\arguments{
\item{CalFluxData}{the CalFluxData object having a platepair object with a loaded reference plate}
}
\description{
Runs quality control on the reference (control) plate read, prior to treatment.
This collects information on parameter variation (standard deviation, and coefficient of variation)
and z-scores for each parameter and each well help to identy outlier/bad wells that we might want to 'mask'
}
