% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{runReferenceQC}
\alias{runReferenceQC}
\title{Runs a quality control process on the reference plate read. The method will export a excel report file.}
\usage{
runReferenceQC(plateSet, rootExportDirectory, plateFile)
}
\arguments{
\item{plateSet}{the plate set to run QC on}

\item{rootExportDirectory}{the directory for report output.}

\item{plateFile}{the plate file name so that the report can capture the input plate file name.}
}
\description{
Runs a quality control process on the reference plate read. The method will export a excel report file.
}
