% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{setParameterAbbreviations}
\alias{setParameterAbbreviations}
\title{Method updates parameter abbreviations to those specified in the parameter file}
\usage{
setParameterAbbreviations(CalFluxData)
}
\arguments{
\item{CalFluxData}{CalFluxData object initialized with parameter file, plate data and plate maps loaded.}
}
\value{
returns the CalFluxData object with parameter abbreviations set to use for all downstream reports.
}
\description{
Method updates parameter abbreviations to those specified in the parameter file
}
