% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{statAnalyis}
\alias{statAnalyis}
\title{performs a statistical analysis}
\usage{
statAnalyis(aso, analysis = "paired-t-test")
}
\arguments{
\item{aso}{an aso object having transformed data}

\item{analysis}{the name of the analysis type, one of c('paired-t-test'), paired-t-test is the default.}
}
\description{
performs a statistical analysis
}
