% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{subsetPlateSet}
\alias{subsetPlateSet}
\title{subsetPlateSet}
\usage{
subsetPlateSet(plateSet, plateIndicesToKeep)
}
\arguments{
\item{plateSet}{The plateset to subset}

\item{plateIndicesToKeep}{The indices with in the plate collection to keep in the new PlateSet object}
}
\description{
Reduce platesSet object to the specified plate indices. This utility method is intended to be run prior to generation of transformed data and QC.
}
