% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{trellisPlateSetViewsPairViews}
\alias{trellisPlateSetViewsPairViews}
\title{Produces a grid of plate QC views for a supplied reference plate, experimental plate
and any transformations applied to the data.}
\usage{
trellisPlateSetViewsPairViews(
  plateSet,
  platePair,
  paramsToPlot,
  showRowColumnLabels = T,
  maxSatCount = 10,
  showLegend = F,
  tMethod = "log2ratio",
  ggplot = T,
  elementDividers = F
)
}
\arguments{
\item{plateSet}{a PlateSet object containing at least one reference and one experimental treatment plate.}

\item{platePair}{a vector of length 2 specifying reference and experimental plate names to QC.}

\item{paramsToPlot}{a vector of parameters that should be visualized.}

\item{showRowColumnLabels}{a boolean to show or hide plate row and column labels. Larger trellises may benefit from hiding labels.}

\item{maxSatCount}{maximumn number of wells to allow color saturation in the plate heatmap. This sets color scale limits such that this number of wells saturate,
evenly at both ends of the plate's value range. This reduces the effect of outliers on the color scale limits.}

\item{showLegend}{a boolean that controls showing or hiding the colorbar legend next to the plate heatmaps.}

\item{tMethod}{is one of c('log2ratio', 'pct_change') and is used to control which data transformation to display}

\item{ggplot}{is a boolean to control if rendering is by ggplot2 or by ComplexHeatmap. Ggplot2 is the default.}

\item{elementDividers}{is a boolean controling whether boarders between heatmap elements are shown or not. Larger trellises benefit from hiding these lines.}
}
\description{
Produces a grid of plate QC views for a supplied reference plate, experimental plate
and any transformations applied to the data.
}
