% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{trellisPlateView}
\alias{trellisPlateView}
\title{Plots plate heatmaps for one plate, for a collection of one or more statistics/parameters}
\usage{
trellisPlateView(
  plate,
  paramsToPlot,
  showRowColumnLabels = T,
  maxSatCount = 10,
  showLegend = F,
  scaleColors = c("#e8eb59", "#0740fa"),
  plateTitle = "",
  showRowTitle = F,
  rowTitle = "row",
  ggplot = T,
  elementDividers = F,
  returnPlots = F
)
}
\arguments{
\item{plate}{plate object with loaded data}

\item{paramsToPlot}{a vector of parameters that should be visualized.}

\item{showRowColumnLabels}{a boolean to show or hide plate row and column labels. Larger trellises may benefit from hiding labels.}

\item{maxSatCount}{maximumn number of wells to allow color saturation in the plate heatmap. This sets color scale limits such that this number of wells saturate,
evenly at both ends of the plate's value range. This reduces the effect of outliers on the color scale limits.}

\item{showLegend}{a boolean that controls showing or hiding the colorbar legend next to the plate heatmaps.}

\item{scaleColors}{a vector of colors to define the color scale, two colors should be provides as low color then high color.}

\item{plateTitle}{a title to show over the heatmap}

\item{showRowTitle}{a boolean to control is the row title is shown.}

\item{rowTitle}{optional row title that can be used to label the plate on the left.}

\item{ggplot}{is a boolean to control if rendering is by ggplot2 or by ComplexHeatmap. Ggplot2 is the default.}

\item{elementDividers}{is a boolean controling whether boarders between heatmap elements are shown or not. Larger trellises benefit from hiding these lines.}

\item{returnPlots}{if true, return a list object of ggplot tile plots for visualization, false will display plots.}
}
\description{
Plots plate heatmaps for one plate, for a collection of one or more statistics/parameters
}
