% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{zFactor}
\alias{zFactor}
\title{Utility analysis method to generate z-prime factor data.}
\usage{
zFactor(FLIPRData, dataType = "transformed")
}
\arguments{
\item{FLIPRData}{FLIPRData object on which to perform the z-prime factor analysis. z-prime = 1- 3*(abs(posCtrlMean-negCtrlMean))/(posCtrlSD + negCtrlSD)}

\item{dataType}{Either transformed or raw data.}
}
\value{
a z-prime factor result table with positive and negative control means, SDs, CVs and the z-prime factors for each parameter.
}
\description{
Utility analysis method to generate z-prime factor data.
}
