% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{addPlate}
\alias{addPlate}
\title{Add a plate to a plateset}
\usage{
addPlate(plateset, plate, label)
}
\arguments{
\item{plateset}{a plateset S4 object}

\item{plate}{an plate S4 object to add to the plateset}
}
\description{
Add a plate to a plateset
}
