% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{applyCategoricalToNumeric}
\alias{applyCategoricalToNumeric}
\title{Converts categorical data to numeric values according to specified rules in the input parameter file.
Note that this method will only run on parameters that are flagged as having categorical data.}
\usage{
applyCategoricalToNumeric(aso)
}
\arguments{
\item{aso}{aso object initialized with parameter file, plate data and plate maps loaded.}
}
\value{
returns the aso object categorical parameters converted to numeric values as specified in the parameter file.
}
\description{
Converts categorical data to numeric values according to specified rules in the input parameter file.
Note that this method will only run on parameters that are flagged as having categorical data.
}
