% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{applyDataCoverageFilters}
\alias{applyDataCoverageFilters}
\title{This method optionally removes parameters that have a large amount of missing data, beyond specified levels.}
\usage{
applyDataCoverageFilters(aso)
}
\arguments{
\item{aso}{aso object initialized with parameter file, plate data and plate maps loaded.}
}
\value{
returns the aso object with plate data reduced to parameters passing a minium data limit as specified in the parameter file.
}
\description{
This method optionally removes parameters that have a large amount of missing data, beyond specified levels.
}
