% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{asoRandomForest}
\alias{asoRandomForest}
\title{generatees random forest predictions for ASOs}
\usage{
asoRandomForest(aso, masking = "")
}
\arguments{
\item{aso}{an aso object containing transformed data that's ready for analysis}

\item{masking}{list of wells to optionally mask}
}
\value{
creates a prediction heatmap and dot plot, returns a Heatmap object.
}
\description{
generatees random forest predictions for ASOs
}
