% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{assessDataCompleteness}
\alias{assessDataCompleteness}
\title{Runs an analysis on the CalFluxData object to report on data coverage for each parameter.
The missing data report is entered into the CalFluxData object.}
\usage{
assessDataCompleteness(CalFluxData)
}
\arguments{
\item{CalFluxData}{CalFluxData object initialized with parameter file, plate data and plate maps loaded.}
}
\value{
returns the CalFluxData object containing the missing data report in the dataCoverageTables slot
}
\description{
Runs an analysis on the CalFluxData object to report on data coverage for each parameter.
The missing data report is entered into the CalFluxData object.
}
