% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{aucParameterPatch}
\alias{aucParameterPatch}
\title{A utility method to replace the input AUC parameter name with a new name.
The original text files contain unicode characters that not not well supported.}
\usage{
aucParameterPatch(FLIPRData)
}
\arguments{
\item{FLIPRData}{the FLIPRData object to patch AUC parameter names}
}
\value{
returns a FLIPRData object with the AUC parameter names patched.
}
\description{
A utility method to replace the input AUC parameter name with a new name.
The original text files contain unicode characters that not not well supported.
}
