% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{calculateParamCV}
\alias{calculateParamCV}
\title{Calculate parameter coefficient of variation}
\usage{
calculateParamCV(plate, plateSize = 384)
}
\arguments{
\item{plate}{FLIPRData plate}

\item{plateSize}{plate size, e.g. 384 (default)}
}
\value{
returns a dataframe with columns 'parameter' and 'CV'
}
\description{
Calculate parameter coefficient of variation
}
