% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{dataTransform}
\alias{dataTransform}
\title{Transforms plate data based on a supplied plate set with a reference.}
\usage{
dataTransform(
  plateSet,
  platePair,
  method = "log2ratio",
  bkgrdCorr = F,
  bkgrdSamples = NULL,
  bkgrdMode = "median",
  firstDataCol = 2
)
}
\arguments{
\item{plateSet}{a plateset with at least one reference and one experimental plate}

\item{method}{a transformation method, one of c('log2ratio', 'pct_control')}

\item{bkgrdCorr}{a boolean that controls if there is subtraction of median background.}
}
\value{
returns a PlateSet object with an additional set of transformed data.
}
\description{
Transforms plate data based on a supplied plate set with a reference.
}
