% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{exportBarCharts}
\alias{exportBarCharts}
\title{This function exports a series of barcharts for each CalFluxData and each parameter into a pdf file.}
\usage{
exportBarCharts(CalFluxData)
}
\arguments{
\item{CalFluxData}{a CalFluxData object having transformed data.}
}
\description{
This function exports a series of barcharts for each CalFluxData and each parameter into a pdf file.
}
