% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{exportBarCharts}
\alias{exportBarCharts}
\title{This function exports a series of barcharts for each ASO and each parameter into a pdf file.}
\usage{
exportBarCharts(aso)
}
\arguments{
\item{aso}{an aso object having transformed data.}
}
\description{
This function exports a series of barcharts for each ASO and each parameter into a pdf file.
}
