% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{exportDataCoverageReport}
\alias{exportDataCoverageReport}
\title{Exports the data coverage report that lists parameters and information on missing data.}
\usage{
exportDataCoverageReport(CalFluxData)
}
\arguments{
\item{CalFluxData}{the CalFluxData object containing pre-computed data coverage results}
}
\description{
Exports the data coverage report that lists parameters and information on missing data.
}
