% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{exportMachineLearningResults}
\alias{exportMachineLearningResults}
\title{Exports a colelction of random forest results into one file, along with Benchmark Dose.}
\usage{
exportMachineLearningResults(FLIPRData)
}
\arguments{
\item{FLIPRData}{The FLIPRData object containing the MachineLearning Results}
}
\description{
Exports a colelction of random forest results into one file, along with Benchmark Dose.
}
