% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{exportPairedTTestResult}
\alias{exportPairedTTestResult}
\title{This function exports a paired t-test result.
 @param FLIPRData a FLIPRData object on which paired ttest has been run
 @param ttestResult a dataframe represenint the ttest result
 @param fileName the t-test result ouput file name.}
\usage{
exportPairedTTestResult(FLIPRData, ttestResult, fileName)
}
\description{
This function exports a paired t-test result.
 @param FLIPRData a FLIPRData object on which paired ttest has been run
 @param ttestResult a dataframe represenint the ttest result
 @param fileName the t-test result ouput file name.
}
