% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{exportPairedTTestResult}
\alias{exportPairedTTestResult}
\title{This function exports a paired t-test result.}
\usage{
exportPairedTTestResult(CalFluxData, ttestResult, fileName)
}
\arguments{
\item{CalFluxData}{a CalFluxData object on which paired ttest has been run}

\item{ttestResult}{a dataframe representing the ttest result}

\item{fileName}{the t-test result output file name.}
}
\description{
This function exports a paired t-test result.
}
