% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{exportPlateViews}
\alias{exportPlateViews}
\title{Exports views of the plate data, for each parameter, as a pdf file.
The exported file will include the date and time of file creation.}
\usage{
exportPlateViews(CalFluxData)
}
\arguments{
\item{CalFluxData}{a CalFluxData object containing plate data and having transformed data}
}
\description{
Exports views of the plate data, for each parameter, as a pdf file.
The exported file will include the date and time of file creation.
}
