% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{exportRfPredictionHeatmaps}
\alias{exportRfPredictionHeatmaps}
\title{Exports prediction heatmaps for each plate read. These are generated based on the FLIPRData classes mlResults slot.}
\usage{
exportRfPredictionHeatmaps(FLIPRData)
}
\arguments{
\item{FLIPRData}{The FLIPRData object containing Random Forest result.}
}
\description{
Exports prediction heatmaps for each plate read. These are generated based on the FLIPRData classes mlResults slot.
}
