% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{exportRfPredictionHeatmaps}
\alias{exportRfPredictionHeatmaps}
\title{Exports prediction heatmaps for each plate read. These are generated based on the aso classes rfResults slot.}
\usage{
exportRfPredictionHeatmaps(aso)
}
\arguments{
\item{aso}{The aso object containing Random Forest result.}
}
\description{
Exports prediction heatmaps for each plate read. These are generated based on the aso classes rfResults slot.
}
