% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{exportTransformedData}
\alias{exportTransformedData}
\title{A utility method to export transformed data. The transformed data represents
change between the ASO treated state and teh control/untreated state.}
\usage{
exportTransformedData(aso)
}
\arguments{
\item{aso}{an aso object from which to export transformed data.}
}
\description{
A utility method to export transformed data. The transformed data represents
change between the ASO treated state and teh control/untreated state.
}
