% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fiterToSelectedParameters}
\alias{fiterToSelectedParameters}
\title{Reduces the plate data object to only contain data for the subset of selected parameters.}
\usage{
fiterToSelectedParameters(aso)
}
\arguments{
\item{aso}{aso object initialized with parameter file, plate data and plate maps loaded.}
}
\value{
returns the aso object with working set of plate data reduced to parameters specified in the parameter file.
}
\description{
Reduces the plate data object to only contain data for the subset of selected parameters.
}
