% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{getBarChartTrellis}
\alias{getBarChartTrellis}
\title{This method returns a trellis or table of bar chart plots for a set of samples and parameters.}
\usage{
getBarChartTrellis(
  plateSet,
  samples,
  parameters,
  samplesIn = "rows",
  tMethod = "log2Ratio"
)
}
\arguments{
\item{plateSet}{the plateset object containing the data.}

\item{samples}{the vector of sample names (compounds, FLIPRData ids), for which to pull data.}

\item{parameters}{a vector of parameters for which to build plots.}

\item{samplesIn}{indicates if samples should be represented as rows or columns.}

\item{tMethod}{the transformation method to export}
}
\value{
returns a grid of bar charts.
}
\description{
This method returns a trellis or table of bar chart plots for a set of samples and parameters.
}
