% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{getParameterCorrelationMatrix}
\alias{getParameterCorrelationMatrix}
\title{creates a correlation plot for paramters}
\usage{
getParameterCorrelationMatrix(aso)
}
\arguments{
\item{aso}{an aso object containing transformed data.}
}
\description{
creates a correlation plot for paramters
}
