% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{getStatisticBarChartFromTransformedPlateSet}
\alias{getStatisticBarChartFromTransformedPlateSet}
\title{This method returns a ggplot object representing the transformed data, for a given sample name and parameter name.}
\usage{
getStatisticBarChartFromTransformedPlateSet(
  plateSet,
  sampleName,
  parameter,
  tMethod = "log2Ratio",
  showTitle = TRUE,
  showXAxisTitle = TRUE,
  showYAxisTitle = TRUE
)
}
\arguments{
\item{plateSet}{the plateset object containing the transformed data}

\item{sampleName}{the sample (compound) to return data for.}

\item{parameter}{the parameter for which to return data.}

\item{tMethod}{the transoformation method for the transformed data.}

\item{showTitle}{whether to show the plot title.}

\item{showXAxisTitle}{whether to show the x-axis title.}

\item{showYAxisTitle}{whether to show the y-axis title.}
}
\description{
This method returns a ggplot object representing the transformed data, for a given sample name and parameter name.
}
