% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{harmonizeParameters}
\alias{harmonizeParameters}
\title{Makes sure that the reference plate read and the treatment plate read have the same parameters
after possibly filtering based on data loss.}
\usage{
harmonizeParameters(aso)
}
\arguments{
\item{aso}{the aso object having plate read pairs for comparison}
}
\description{
Makes sure that the reference plate read and the treatment plate read have the same parameters
after possibly filtering based on data loss.
}
