% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{loadPlateMaps}
\alias{loadPlateMaps}
\title{This function loads plate annotations, plate maps as specified in the aso object after initializing with parameter file}
\usage{
loadPlateMaps(aso)
}
\arguments{
\item{aso}{the input ASO file generated by readAsoParameterFile}
}
\value{
returns an aso object containing a plateset object with loaded plate data and plate map annotations.
}
\description{
This function loads plate annotations, plate maps as specified in the aso object after initializing with parameter file
}
