% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{loadPlateMaps}
\alias{loadPlateMaps}
\title{This function loads plate annotations, plate maps as specified in the CalFluxData object after initializing with parameter file}
\usage{
loadPlateMaps(CalFluxData)
}
\arguments{
\item{CalFluxData}{the input CalFluxData file generated by readCalFluxDataParameterFile}
}
\value{
returns a CalFluxData object containing a plateset object with loaded plate data and plate map annotations.
}
\description{
This function loads plate annotations, plate maps as specified in the CalFluxData object after initializing with parameter file
}
