% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{loadPlates}
\alias{loadPlates}
\title{This method loads plate data according the the plate files specified within the FLIPRData parameter file.
One plate object is created for each input plate file and the plate pairs are stored within the returned FLIPRData object.}
\usage{
loadPlates(FLIPRData)
}
\arguments{
\item{FLIPRData}{the input FLIPRData file generated by readFLIPRDataParameterFile}
}
\value{
returns a FLIPRData object containing a plateset object with loaded plate data.
}
\description{
This method loads plate data according the the plate files specified within the FLIPRData parameter file.
One plate object is created for each input plate file and the plate pairs are stored within the returned FLIPRData object.
}
