% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{maskEmptyWells}
\alias{maskEmptyWells}
\title{Converts well vales to NaN for wells that are annotated in the platemap as "Empty".
Downstream methods will not be impacted by these well values.}
\usage{
maskEmptyWells(aso)
}
\value{
returns the aso object with specified wells masked.
}
\description{
Converts well vales to NaN for wells that are annotated in the platemap as "Empty".
Downstream methods will not be impacted by these well values.
}
