% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot PCA results}
\usage{
plotPCA(CalFluxData, pcaList, pcaType = 1, plotLoadings = F, dataName = "")
}
\arguments{
\item{CalFluxData}{The CalFluxData object containing pcaResults}

\item{pcaList}{A list object containing PCA results}

\item{pcaType}{1 = wells pca, 2 = parameters pca}

\item{plotLoadings}{a boolean value determining if PCA loadings should be plotted.}

\item{dataName}{an optional string to tag on the data results.}
}
\description{
Plot PCA results
}
