% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{rawPlateMatrixStatsToPlateFormat}
\alias{rawPlateMatrixStatsToPlateFormat}
\title{rawPlateMatrixStatsToPlateFormat takes a plate object and returns a list of matrix objects, each formatted
according to plate dimensions. The list of matrix objects is named according to a supplied vector of measured parameters}
\usage{
rawPlateMatrixStatsToPlateFormat(plate, parameters)
}
\arguments{
\item{plate}{The plate containing the data}

\item{paramaters}{the vector of measured parameter to report on}
}
\value{
Returns a list of plate matrix objects, named by each of the measured parameters in parameter
}
\description{
rawPlateMatrixStatsToPlateFormat takes a plate object and returns a list of matrix objects, each formatted
according to plate dimensions. The list of matrix objects is named according to a supplied vector of measured parameters
}
