% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_methods.R
\name{rawPlateMatrixStatToPlateFormat}
\alias{rawPlateMatrixStatToPlateFormat}
\title{rawPlateMatrixStatToPlateFormat takes a plate object and returns a matrix formatted
according to plate dimensions containing data for a specific measured parameter}
\usage{
rawPlateMatrixStatToPlateFormat(plate, parameter)
}
\arguments{
\item{plate}{The plate containing the data}

\item{parameter}{the measured parameter to report on}
}
\value{
returns a matrix with dimensions corresponding to plate type. A01 is in upper left.
}
\description{
rawPlateMatrixStatToPlateFormat takes a plate object and returns a matrix formatted
according to plate dimensions containing data for a specific measured parameter
}
