% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{replaceBlanksAndNAValues}
\alias{replaceBlanksAndNAValues}
\title{Replaces 'Blank' or Empty data and N/A data for each parameter according to specified rules for each parameter.}
\usage{
replaceBlanksAndNAValues(aso)
}
\arguments{
\item{aso}{aso object initialized with parameter file, plate data and plate maps loaded.}
}
\value{
returns the aso object categorical parameters converted to numeric values as specified in the parameter file.
}
\description{
Replaces 'Blank' or Empty data and N/A data for each parameter according to specified rules for each parameter.
}
