% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{runTTest}
\alias{runTTest}
\title{Generic T test function for multiple experimental designs}
\usage{
runTTest(CalFluxData)
}
\arguments{
\item{CalFluxData}{CalFluxData object having loaded plate data.}
}
\value{
a dataframe containing stat results for each parameter, for each well.
}
\description{
Generic T test function for multiple experimental designs
}
