% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{setParameterAbbreviaions}
\alias{setParameterAbbreviaions}
\title{Method updates parameter abbreviations to those specified in the parameter file}
\usage{
setParameterAbbreviaions(FLIPRData)
}
\arguments{
\item{FLIPRData}{FLIPRData object initialized with parameter file, plate data and plate maps loaded.}
}
\value{
returns the FLIPRData object with parameter abbreviations set to use for all downstream reports.
}
\description{
Method updates parameter abbreviations to those specified in the parameter file
}
