% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{transformData}
\alias{transformData}
\title{Transforms the plate data, currently taking a log base 2 transformation of the ratio of the data
This produces a single table that reports on the comparison between the reference or control state
versus the aso treated state.}
\usage{
transformData(aso)
}
\arguments{
\item{aso}{an aso object having plate data (a platepair object) to transform}
}
\value{
an aso object with platepair object now containing the transformed data
}
\description{
Transforms the plate data, currently taking a log base 2 transformation of the ratio of the data
This produces a single table that reports on the comparison between the reference or control state
versus the aso treated state.
}
