% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{transformData}
\alias{transformData}
\title{Transforms the plate data using the configured transformation method
This produces a single table that reports on the comparison between the reference or control state
versus the CalFluxData treated state.}
\usage{
transformData(CalFluxData, paired_analysis)
}
\arguments{
\item{CalFluxData}{a CalFluxData object having plate data (a platepair object) to transform}

\item{paired_analysis}{boolean specifying if plate pairs are used for baseline correction}
}
\value{
a CalFluxData object with platepair object now containing the transformed data
}
\description{
Transforms the plate data using the configured transformation method
This produces a single table that reports on the comparison between the reference or control state
versus the CalFluxData treated state.
}
